Store image from external
This is a one line code I am testing to see if the image can be stored only in the folder on the server (since everyone else will need access) and the path in the field of the table it belongs to. It's almost the same concept as you showed me. I am testing your method.
Thank you Hassan for your help